refactor(buffer): phase 2D — move BufferFormat + helpers to format.rs
Extracts the BufferFormat struct and its impl, the LineEnding enum
and its impl, and the seven detection/conversion free fns
(detect_line_ending, detect_encoding, detect_encoding_or_binary,
detect_and_convert_encoding, convert_to_encoding,
normalize_line_endings, convert_line_endings_to) from mod.rs into
model/buffer/format.rs as a proper module.

mod.rs now has `pub mod format;` and re-exports BufferFormat and
LineEnding for API compatibility. The seven helpers — which were
previously static methods on `impl TextBuffer` (Self::detect_*) —
are now free pub fns; the in-file call sites (primarily
build_write_recipe) now invoke them via `format::detect_*`. The
inline test `mod tests` sites rewrite `TextBuffer::detect_*` to
`super::format::detect_*`.

convert_line_endings_to remains pub(super) — only the save path and
its tests use it, and it's not API surface.

Buffer test suite (127 tests) green.
